Move socket-related code to socket.rs
