+++ src/lib.rs
@@ -78,6 +78,7 @@ mod debug;
 mod system;
 mod traits;
 mod utils;
+mod socket;
 
 /// This function is only used on linux targets, on the other platforms it does nothing and returns
 /// `false`.
